Change the TRACE macro to use std::format instead of snprintf (#24)
C++20 comes with a replacement for printf, use it to print debug strings.

Note 1: To allow compile time  error detection the call can not be forwarded, but the macro must be expanded in place.

Note 2: there is no simple way to print the address of a pointers, the standard way is to do a static_cast to void*.
In essence this makes sense as std::format is designed to auto detect the type and be type safe. It is however more boilerplate.