Add COM support for IPropertyStore
Explorer uses IPropertyStore to retrieve additional meta data to display in the preview pane, info tooltip and property tab sheet.
As the Netpbm format is unknown to Windows, the default photo property handler cannot be used and a custom implementation is needed.

This property store provides the following standard properties:
- System.Image.HorizontalSize
- System.Image.VerticalSize
- System.Image.BitDepth
- System.Image.Dimensions
- System.Image.Compression